Refactor some code Add Host to log output now we support multiple incoming hostnames
