+++ internal/config/util.go
@@ -1,81 +1,11 @@
 package config
 
 import (
-	"fmt"
-	"log"
 	"math"
-	"net/http"
-	"net/url"
 	"strconv"
 	"strings"
 	"time"
 )
-
-const xForwardedProto = "X-Forwarded-Proto"
-
-// FormatBaseAddress returns the public base address of the server given the server config
-// it ignores the port if it's default for the given protocol
-func (c *Config) FormatBaseAddress(r *http.Request) string {
-
-	scheme := "http"
-	host := c.Address
-
-	// We want to use all Host-sourced configs, or none, so check the Host header is non-empty too.
-	if r != nil && r.Host != "" && r.Header.Get(xForwardedProto) != "" {
-		scheme = r.Header.Get(xForwardedProto)
-	} else if c.ProtoTLS {
-		scheme = "https"
-	} else if r != nil && r.URL != nil && r.URL.Scheme != "" {
-		scheme = r.URL.Scheme
-	}
-
-	if r != nil && r.Host != "" { // Host headers override all local config
-		host = r.Host
-	} else if c.Port > 0 && ((c.ProtoTLS && c.Port != 443) || (!c.ProtoTLS && c.Port != 80)) { // else config settings apply for non-standard ports
-		host = fmt.Sprintf("%s:%d", c.Address, c.Port)
-	}
-
-	url, err := url.Parse(fmt.Sprintf("%s://%s/", scheme, host))
-	if err != nil {
-		log.Println("failed to parse URL:", err)
-		return ""
-	}
-
-	return url.String()
-}
-
-func (c *Config) FormatBaseAddressStringBuilder(r *http.Request) string {
-
-	var b strings.Builder
-
-	// We want to use all Host-sourced configs, or none, so check the Host header is non-empty too.
-	if r != nil && r.Host != "" && r.Header.Get(xForwardedProto) != "" {
-		b.WriteString(r.Header.Get(xForwardedProto))
-	} else if c.ProtoTLS {
-		b.WriteString("https")
-	} else if r != nil && r.URL != nil && r.URL.Scheme != "" {
-		b.WriteString(r.URL.Scheme)
-	} else {
-		b.WriteString("http")
-	}
-
-	if r != nil && r.Host != "" { // Host headers override all local config
-		b.WriteString("://")
-		b.WriteString(r.Host)
-	} else if c.Port > 0 && ((c.ProtoTLS && c.Port != 443) || (!c.ProtoTLS && c.Port != 80)) { // else config settings apply for non-standard ports
-		b.WriteString("://")
-		b.WriteString(c.Address)
-		b.WriteRune(':')
-		b.WriteString(strconv.Itoa(c.Port))
-	} else {
-		b.WriteString("://")
-		b.WriteString(c.Address)
-	}
-
-	b.WriteRune('/')
-
-	return b.String()
-}
 
 type timeString struct {
 	b strings.Builder
